+++ src/cppSocketLib.cpp
@@ -1,7 +1,23 @@
+/*
+ * Socket Library - cppSocketWrapper
+ * Copyright (C) 2024, Operating Systems II.
+ * Apr 23, 2024.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ */
+
 #include "cppSocketLib.hpp"
 
 IConnection::IConnection(const std::string& address, const std::string& port, bool isBlocking)
-    : address_(address), m_port_(port), isBlocking_(isBlocking)
+    : m_address(address), m_port(port), m_isBlocking(isBlocking)
 {
 }
