--- src/main.cpp
@@ -1,8 +1,0 @@
-#include "cppSocketLib.hpp"
-
-int main()
-{
-    
-
-    return 0;
-}
