--- tests/unit/test1.cpp
@@ -1,60 +1,0 @@
-#include "cppSocketLib.hpp"
-#include <gtest/gtest.h>
-
-TEST(UDPConnectionTestIPv4, BindSuccess) {
-
-    std::shared_ptr<IConnection> con = createConnection("127.0.0.1", "65536", false, UDP);
-    EXPECT_TRUE(con->bind());
-}
-
-TEST(UDPConnectionTestIPv4, CreateFailure) {
-
-    try
-    {
-        std::shared_ptr<IConnection> con1 = createConnection("127.0.0.0.1", "65536", false, UDP);
-    }
-    catch(const std::exception& e)
-    {
-        EXPECT_THROW(throw std::runtime_error("Error getting address"), std::runtime_error);
-    }
-    
-}
-
-TEST(UDPConnectionTestIPv6, BindSuccess) {
-
-    std::shared_ptr<IConnection> con = createConnection("::1", "65536", false, UDP);
-    EXPECT_TRUE(con->bind());
-}
-
-TEST(UDPConnectionTestIPv6, CreateFailure) {
-
-    try
-    {
-        std::shared_ptr<IConnection> con1 = createConnection(":::1", "65536", false, UDP);
-    }
-    catch(const std::exception& e)
-    {
-        EXPECT_THROW(throw std::runtime_error("Error getting address"), std::runtime_error);
-    }
-    
-}
-
-TEST(UDPConnectionTestIPv4, UnsupportedProtocolMacro){
-    EXPECT_THROW(createConnection("127.0.0.1", "8080", true, 10), std::invalid_argument);
-
-}
-
-
-TEST(UDPConnectionTestIPv4, ConnectSuccess) {
-    
-}
-
-TEST(UDPConnectionTestIPv4, ConnectFailure) {
-    
-}
-
-int main(int argc, char** argv)
-{
-    ::testing::InitGoogleTest(&argc, argv);
-    return RUN_ALL_TESTS();
-}
